Add some ergonomic helpers

+++ src/headers/typed/cseq.rs
@@ -8,12 +8,12 @@ use std::convert::TryFrom;
 /// The `CSeq` header in its [typed](super) form.
 #[derive(TypedHeader, Eq, PartialEq, Clone, Debug)]
 pub struct CSeq {
-    pub seq: u16,
+    pub seq: u32,
     pub method: Method,
 }
 
-impl From<(u16, Method)> for CSeq {
-    fn from(tuple: (u16, Method)) -> Self {
+impl From<(u32, Method)> for CSeq {
+    fn from(tuple: (u32, Method)) -> Self {
         Self {
             seq: tuple.0,
             method: tuple.1,
@@ -26,7 +26,7 @@ impl<'a> TryFrom<Tokenizer<'a>> for CSeq {
 
     fn try_from(tokenizer: Tokenizer) -> Result<Self, Self::Error> {
         Ok(CSeq {
-            seq: tokenizer.seq.parse::<u16>()?,
+            seq: tokenizer.seq.parse::<u32>()?,
             method: tokenizer.method.parse::<Method>()?,
         })
     }
